feat(plugins): add Plugin Registry with manifests, accordion settings UI, and AI Provider extraction
