+++ packages/plugins/src/index.ts
@@ -1,1 +1,7 @@
 export { ALL_PLUGINS } from "./plugin-info";
+export { PLUGIN_MANIFESTS } from "./registry";
+export type {
+	PluginCommandDef,
+	PluginManifest,
+	PluginSettingsProps,
+} from "./types";
